% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nls_tiles.R
\name{nls_download_tile_png}
\alias{nls_download_tile_png}
\title{Download a single NLS town-plan PNG tile}
\usage{
nls_download_tile_png(
  base_url = "https://mapseries-tilesets.s3.amazonaws.com/os/town-england/North",
  z,
  x,
  y,
  destfile
)
}
\arguments{
\item{base_url}{Base URL for the NLS tile set.}

\item{z}{Zoom level.}

\item{x}{Tile x coordinate.}

\item{y}{Tile y coordinate.}

\item{destfile}{Destination file path (a \code{.png}).}
}
\value{
\code{TRUE} if the file exists after download; otherwise \code{FALSE}.
}
\description{
Download a single NLS town-plan PNG tile
}
