% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nls_tiles.R
\name{nls_tile_rast_lonlat}
\alias{nls_tile_rast_lonlat}
\title{Read an NLS tile by lon/lat}
\usage{
nls_tile_rast_lonlat(
  base_url = "https://mapseries-tilesets.s3.amazonaws.com/os/town-england/North",
  z,
  lon,
  lat
)
}
\arguments{
\item{base_url}{Base URL for the NLS tile set.}

\item{z}{Zoom level.}

\item{lon}{Longitude.}

\item{lat}{Latitude.}
}
\value{
A \link[terra:SpatRaster-class]{terra::SpatRaster} in EPSG:3857.
}
\description{
Convenience wrapper around \code{tile_coords_to_xy()} and \code{nls_tile_rast_zxy()}.
}
