% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_xyz.R
\name{tile_webmerc_extent}
\alias{tile_webmerc_extent}
\title{Web Mercator extent of an XYZ tile}
\usage{
tile_webmerc_extent(z, x, y)
}
\arguments{
\item{z}{Zoom level.}

\item{x}{Tile x coordinate.}

\item{y}{Tile y coordinate.}
}
\value{
A \code{\link[terra:ext]{terra::ext()}} object.
}
\description{
Web Mercator extent of an XYZ tile
}
